Implement install command
The command installs the given ruby version as RPM, sets symbolic
links in ~/bin and adjusts GEM_HOME and PATH so that the right
binaries are picked up and gems are stored locally in ~/.gems/rubyx.y.
It also sets the .gemrc install option to not create ruby version
suffices so that the standard executables are generated and picked up
from the .gems directory.